+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+- 2024-08-09:
+  - added search export endpoint.
+  - refactored search endpoint as GET.
 - 2024-08-05: made `ITokenFilter` async.
 - 2024-07-31: ⚠️ model refactoring in progress, bumping all the versions to 4.x.x (release 3.0.3 is the last with the old model): a more denormalized storage model replaces the original one to enhance performance and simplify code:
   - the SQL schema has been updated to reflect the new model.